Line wrappings and general code cleanups.
While I'm not entirely satisfied with how I've currently wrapped the lines it's
a lot better (in my opinion) than before.

Besides wrapping lines I've also performed general code cleanups to keep jslint
as happy as possible. For example, the single line if/for statements (as well
as the crazy switch/case statements) have been replaced with multi line
statements.

While this greatly increases the amount of lines used it also makes it a lot
easier to read and keeps things consistent. If the amount of lines is important
then it's time to start using a minifier.

Last but not least I've also changed a few variable declarations to again keep
jslint happy as well as a preparation for documenting the various variables and
functions in following commits.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>